pyproject.toml
[project]
name = "yearprogressbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "apscheduler>=3.11.1",
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
]
